move GM5 polygon functions to GM5 compat extension
